requirements.txt
numpy
pandas
openpyxl
pyodbc
plotly
pypdf2
reportlab
sentencepiece
dateparser
torch
textblob
pyyaml